Add zero-copy WriteBuffer and route write() through it
- Write via IORING_OP_WRITE_FIXED without copying:
  write_buffer() acquires a slot and returns an owned
  WriteBuffer backed by the registered slab; the caller
  fills it through as_mut(), records the length with
  set_len(), and hands it to write(), which submits the
  slot pointer directly.
- WriteBuffer stores only { ptr, len, bid, generation };
  capacity() is the pool's slot size resolved through the
  TLS runtime pointer (generation-guarded), so the whole
  slot is usable and no oversized-buffer EINVAL check is
  needed (set_len asserts the capacity instead).
- Drop owns recycling: write() no longer copies the
  payload, records the slot id, or releases the slot
  itself; the buffer recycles on drop via the thread-local
  runtime pointer and skips the recycle after runtime
  shutdown, like ReadBuffer.
- Exposed at the crate root as coio::WriteBuffer.
- Narrow ReadBuffer.len to u32.
- Tests: WriteBuffer drop/stale-generation/capacity/
  as_mut-set_len-clear/set_len-panic unit tests and
  block_on_write_to_file/write_full_slot/
  write_many_recycles integration tests.

+++ libs/coio/src/pbuf.rs
@@ -206,13 +206,13 @@ impl Drop for ProvidedBufferPool {
 /// the runtime is gone.
 pub struct ReadBuffer {
     ptr: NonNull<u8>,
-    len: usize,
+    len: u32,
     bid: u16,
     generation: u64,
 }
 
 impl ReadBuffer {
-    pub(crate) fn new(ptr: NonNull<u8>, len: usize, bid: u16, generation: u64) -> Self {
+    pub(crate) fn new(ptr: NonNull<u8>, len: u32, bid: u16, generation: u64) -> Self {
         Self {
             ptr,
             len,
@@ -223,7 +223,8 @@ impl ReadBuffer {
 
     /// Copies the buffer into an owned `Vec` and recycles the pool slot.
     pub fn into_vec(mut self) -> Vec<u8> {
-        let data = unsafe { core::slice::from_raw_parts(self.ptr.as_ptr(), self.len) }.to_vec();
+        let data =
+            unsafe { core::slice::from_raw_parts(self.ptr.as_ptr(), self.len as usize) }.to_vec();
         self.recycle();
         core::mem::forget(self);
         data
@@ -246,7 +247,7 @@ impl Deref for ReadBuffer {
     type Target = [u8];
 
     fn deref(&self) -> &[u8] {
-        unsafe { core::slice::from_raw_parts(self.ptr.as_ptr(), self.len) }
+        unsafe { core::slice::from_raw_parts(self.ptr.as_ptr(), self.len as usize) }
     }
 }
